changed logic for timeinput validation, bug fixes
changed input validation so that it checks state instead of event.target.value in order to update the state. Now users no longer have to backspace the ":" when inputting a time it will be done automatically. Fixed a bug where the the last two characters of the timeinputs could be non numbers.