bug fixes to timeinputs, changing store Times state when changing settings
Fixed a bug in checkValidInput(now named checkValidInputAuto) where a colon could be entered in as the third character. the states in the store will change when ever the user changes the time settings to match, right now only military time can be inputted and work because passing down AM and PM to the store has not been implemented.