Extra Validation Layer, updated reducer, bug fixes to previnput
Added an extra validation layer to the TimeInput so that the component won't add a invalid time to the store. Activated when saved time button is pressed and gives an error message above the inputs based on local state invalidInput. Updated reducer to take into account AM and PM when settings are set to regular time. Fixed bug where prevInput wasn't being updated when the store changed.