Refactored Inputs, Added new singular input component
Refactored Inputs so that it uses a new component (Input.js) twice and passes it's states down to them. This fixes a bug with colonChecker and being able to input a colon in one input if the other one already had it. Also makes code more lean and readable. Button1 is also now associated with sleep time instead of waketime instead of button2.