score calculation complete(primary)
Implemented score calculation updating using 3 functions that calculate a different part of the score individually and then adding them up. slight refacting of said code for more readable and usability(a helper function to convert time into a number in particular). Only the preliminary stages atm will iterate on it later to be more dynamic.