Separated TimeOptions, added setting tab, changed history and reducer for totalhours
Separated TimeOptions into two new components Settings and CurrentTime. Added a link into the header in order to access the settings (you have to go to this tab to change date format and time settings now). Added totalHours properties to the reducer to display in history, also added function into reducer for the calculation of this property.