connection to backend complete
Connection to beckend complete the website can now synch up with data within the data base. Had to rewrite some of the signup logic in order to introduce a new state that will help with the put request used in synching the store state with the online data in firebase.