persistent login added
Added persistant login, users already logged in when they leave the site or refresh it will automatically be logged in when they come back. After one hour users will have to login again in order to refresh the expire time. edits to Login.js, App.js and the reducer made to do this.